requirements.txt
langchain
python-dotenv
langchain-google-genai
langchain-experimental
qdrant-client
redis
langchain-community
langchain-ollama
pydantic
fastapi
docling
docling-core
uvicorn
img2table
easyocr
hf_xet
google-api-python-client
rank-bm25
sentence-transformers
langgraph
python-multipart
requests
pypdf
